Moved PrereqOrder criterion implementation
